+++ libs/ngx-localstorage/src/index.ts
@@ -1,9 +1,19 @@
 /*
  * Public API Surface of ngx-localstorage
  */
-
+// module
+export * from './lib/ngx-localstorage.module';
+// services
 export * from './lib/services/ngx-localstorage.service';
 export * from './lib/services/storage-event.service';
+// directives
 export * from './lib/directives/ngx-localstorage.directive';
+// decorator
 export * from './lib/decorators';
-export * from './lib/ngx-localstorage.module';
+// interfaces
+export * from './lib/interfaces/decorator-options';
+export * from './lib/interfaces/storage-configuration';
+export * from './lib/interfaces/storage-serializer';
+// tokens
+export * from './lib/tokens/storage-config';
+export * from './lib/tokens/storage-serializer';