+++ libs/ngx-localstorage/src/lib/ngx-localstorage.module.ts
@@ -1,9 +1,14 @@
 import { ModuleWithProviders, NgModule, Optional, SkipSelf } from '@angular/core';
 
-import { NgxLocalstorageConfiguration } from './interfaces';
+import { NgxLocalstorageConfiguration } from './interfaces/storage-configuration';
 import { LocalStorageDirective } from './directives/ngx-localstorage.directive';
-import { NGX_LOCAL_STORAGE_CONFIG } from './token';
+import { NGX_LOCAL_STORAGE_CONFIG } from './tokens/storage-config';
+import { DefaultSerializer } from './classes/default-serializer';
+import { NGX_LOCAL_STORAGE_SERIALIZER } from './tokens/storage-serializer';
 
+/**
+ * Provides the librarys module.
+ */
 @NgModule({
   imports: [
   ],
@@ -12,10 +17,20 @@ import { NGX_LOCAL_STORAGE_CONFIG } from './token';
   ],
   exports: [
     LocalStorageDirective
+  ],
+  providers: [
+    {
+      provide: NGX_LOCAL_STORAGE_SERIALIZER,
+      useClass: DefaultSerializer
+    }
   ]
 })
 export class NgxLocalStorageModule {
-  static forRoot(config?: NgxLocalstorageConfiguration): ModuleWithProviders<NgxLocalStorageModule> {
+
+  /**
+   * Initializes the module for the root module with the given configuration.
+   */
+  public static forRoot(config?: NgxLocalstorageConfiguration): ModuleWithProviders<NgxLocalStorageModule> {
     return {
       ngModule: NgxLocalStorageModule,
       providers: [
@@ -27,6 +42,9 @@ export class NgxLocalStorageModule {
     };
   }
 
+  /**
+   * Creates a new instance.
+   */
   constructor(@Optional() @SkipSelf() parentModule: NgxLocalStorageModule) {
     if (parentModule) {
       throw new Error('NgxLocalStorageModule is already loaded. Import it in the AppModule only');